Moves assets into better folder names. Resolves path issues. Adds level1. Adds a good amount of output for the state of the GameState. Adds MainMenu return from the game state. Changes MainMenu to allow loading Level1 and TrainingLevel(with text update).

+++ unreal-project/GreyMatter/Source/GreyMatter/GreyMatterGameMode.h
@@ -10,7 +10,7 @@ class AGreyMatterGameMode : public AGameMode
 	GENERATED_BODY()
 
 	float playerAmmo;
-	float remainingTime;
+	int32 remainingTime;
 	float score;
 
 public:
